v0.0.7-SNAPSHOT: - Implementation of the relationship between Like, Shop and User has been finished - Refactored attributes and method names from the Like.java class; - Removed ExistsUserIdException.java class and created two exceptions: UserNotLikeRecipeException.java and UserLikedRecipeException.java classes; - Rename methods from RecipeController.java that like and dislike a recipe;  - Was make any improvements in the integration between Recipe, like and user;  - The check of whether the logged-in user likes a recipe when adding or removing a like has been moved to the service layer.

+++ src/main/java/com/recipe/book/api/services/UserServiceImpl.java
@@ -1,6 +1,8 @@
 package com.recipe.book.api.services;
 
 import com.recipe.book.api.exceptions.InvalidCredentialsException;
+import com.recipe.book.api.exceptions.UserLikedRecipeException;
+import com.recipe.book.api.exceptions.UserNotLikeRecipeException;
 import com.recipe.book.api.model.User;
 import com.recipe.book.api.repositories.UserRepository;
 import io.javalin.security.BasicAuthCredentials;
@@ -61,11 +63,17 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void addRecipeToListLike(User currentUser, String recipeId) {
+        if (currentUser.getLikedRecipesList().contains(recipeId)) {
+            throw new UserLikedRecipeException("User id: " + currentUser.getId() + " already liked the recipe!");
+        }
         repository.addRecipeToListLike(currentUser, recipeId);
     }
 
     @Override
     public void removeRecipeFromListLike(User currentUser, String recipeId) {
+        if (!currentUser.getLikedRecipesList().contains(recipeId)) {
+            throw new UserNotLikeRecipeException("User id: " + currentUser.getId() + " did not like the recipe!");
+        }
         repository.removeRecipeFromListLike(currentUser, recipeId);
     }
 }
